+++ measurement/plugins/ip_route/results.py
@@ -7,5 +7,5 @@ if six.PY3 and not sys.version_info.minor == 5:  # All python 3 expect for 3.5
     from .results_py3 import *
 else:
     IPRouteMeasurementResult = collections.namedtuple(
-        "IPRouteMeasurementResult", "id errors host hop_count ip trace",
+        "IPRouteMeasurementResult", "id errors host hop_count ip route",
     )